move ExitLinkActionEvent to analytics module
this action is now used by more than the tool renderer, so it should be in a more centralized place
