+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/ButtonViewHolder.java
@@ -6,7 +6,7 @@ import android.net.Uri;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import org.cru.godtools.base.tool.analytics.model.ExitLinkActionEvent;
+import org.cru.godtools.analytics.model.ExitLinkActionEvent;
 import org.cru.godtools.base.tool.model.view.TextViewUtils;
 import org.cru.godtools.base.ui.util.WebUrlLauncher;
 import org.cru.godtools.tract.R;
